+++ packages/components/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [35.5.2](https://github.com/transferwise/neptune-web/compare/@transferwise/components@35.5.1...@transferwise/components@35.5.2) (2021-04-19)
+
+**Note:** Version bump only for package @transferwise/components
+
+
+
+
+
+## [35.5.1](https://github.com/transferwise/neptune-web/compare/@transferwise/components@35.5.0...@transferwise/components@35.5.1) (2021-04-16)
+
+
+### Bug Fixes
+
+* using requestAnimationFrame to perform the calculation before the next repaint ([#899](https://github.com/transferwise/neptune-web/issues/899)) ([d196340](https://github.com/transferwise/neptune-web/commit/d196340cf12461fe5d628318042bcb0a609c8db1))
+
+
+
+
+
+# [35.5.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@35.4.0...@transferwise/components@35.5.0) (2021-04-15)
+
+
+### Features
+
+* add initials and badge logic to AvatarWrapper ([#896](https://github.com/transferwise/neptune-web/issues/896)) ([aa895e3](https://github.com/transferwise/neptune-web/commit/aa895e324dd3039e50deed4d381567da73ef426a))
+
+
+
+
+
 # [35.4.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@35.3.3...@transferwise/components@35.4.0) (2021-04-08)